Struct circular::Buffer

pub struct Buffer { /* fields omitted */ }

the Buffer contains the underlying memory and data positions

In all cases, 0 ≤ position ≤ end ≤ capacity should be true

Methods

impl Buffer

fn with_capacity(capacity: usize) -> Buffer

allocates a new buffer of maximum size capacity

fn from_slice(data: &[u8]) -> Buffer

allocates a new buffer containing the slice data

the buffer starts full, its available data size is exactly data.len()

fn grow(&mut self, new_size: usize) -> bool

increases the size of the buffer

this does nothing if the buffer is already large enough

fn available_data(&self) -> usize

returns how much data can be read from the buffer

fn available_space(&self) -> usize

returns how much free space is available to write to

fn capacity(&self) -> usize

returns the underlying vector's size

fn empty(&self) -> bool

returns true if there is no more data to read

fn consume(&mut self, count: usize) -> usize

advances the position tracker

if the position gets past the buffer's half, this will call shift() to move the remaining data to the beginning of the buffer

fn fill(&mut self, count: usize) -> usize

after having written data to the buffer, use this function to indicate how many bytes were written

if there is not enough available space, this function can call shift() to move the remaining data to the beginning of the buffer

fn position(&self) -> usize

Get the current position

Examples

use circular::Buffer;
use std::io::{Read,Write};

let mut output = [0;5];

let mut b = Buffer::with_capacity(10);

let res = b.write(&b"abcdefgh"[..]);

b.read(&mut output);

// Position must be 5
assert_eq!(b.position(), 5);
assert_eq!(b.available_data(), 3);

fn reset(&mut self)

moves the position and end trackers to the beginning this function does not modify the data

fn data(&self) -> &[u8]

returns a slice with all the available data

fn space(&mut self) -> &mut [u8]

returns a mutable slice with all the available space to write to

fn shift(&mut self)

moves the data at the beginning of the buffer

if the position was more than 0, it is now 0